requirements.txt
mcp==0.1.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
